+++ server/app/controllers/PublisherController.js
@@ -11,7 +11,7 @@ exports.getAll = async function (req, res, next) {
 
 exports.getAllByAdmin = async function (req, res, next) {
     const authData = await jwtHelper.decodeToken(req, res);
-    if (authData !== null) {
+    if (authData !== null && authData.role === "admin") {
         let items = await publisherDb.getAll();
         for (item of items) {
             let user = await userDb.getById(item.userId);
@@ -79,9 +79,14 @@ exports.deleteById = async function (req, res, next) {
         if (error) {
             return res.status(400).send({error})
         }
-        let deletedItem = await publisherDb.deletePublisherById(itemId);
-        await logHelper.createLog(deletedItem.name + ' deleted.', email, "publisher-crud");
-        return res.status(200).send({message: deletedItem.name + ' deleted'});
+        let existingItem = await publisherDb.getById(itemId);
+        if(existingItem) {
+            await publisherDb.deletePublisherById(itemId);
+            await logHelper.createLog(existingItem.name + ' deleted.', email, "publisher-crud");
+            return res.sendStatus(200);
+        }else {
+            return res.status(404).send({error: "Console not found"});
+        }
     }
     if (authData !== null && authData.role === "user") {
         let error = publisherValidation.validateDelete(req);
@@ -90,11 +95,11 @@ exports.deleteById = async function (req, res, next) {
         }
         let user = await userDb.getUser(email);
         if (user) {
-            let ownGame = await publisherDb.getPublisherByUserandId(user._id, itemId);
-            if (ownGame) {
-                let deletedItem = await publisherDb.deletePublisherById(itemId);
-                await logHelper.createLog(deletedItem.name + ' deleted.', email, "publisher-crud");
-                return res.status(200).send({message: deletedItem.name + ' deleted'});
+            let ownItem = await publisherDb.getPublisherByUserandId(user._id, itemId);
+            if (ownItem) {
+                await publisherDb.deletePublisherById(itemId);
+                await logHelper.createLog(ownItem.name + ' deleted.', email, "publisher-crud");
+                return res.status(200).send({message: ownItem.name + ' deleted'});
             } else {
                 return res.sendStatus(403);
             }
